Make create-product cURL script and test it
Made a create product cURL script and tested it. Was running into a
decent amount of 404 errors, but team found out that we were missing
an assignment line in products.js controller so we added the line
and now we are able to post to the server successfully.
Next Steps: Test index and show of Products with cURL scripts