Create cURL script to test creating PastOrder
Team created a cURL script to test adding a PastOrder to the database.
Used the script to add two entries to the PastOrder collection, one with
a comment specified and one without to test default; both are working.
Checked the state of the database and everything is good to go.
NextSteps: Test index, update, and destroy past order with cURL scripts.