Refactor users#updateCart to push one req object
As per the new way to update the cart on a user, refactored the old
updateCart function to take a single cart product item and push it onto
the cart array of the user. If successful, the response returns the
entire user object so front-end data can be updated.
NOTE: Would like to just return the updated cart in the response, but
I get a warning of 'user.cart' being undefined. Will have to figure out.
Next Steps: Finish front-end request chain and test.