Modify user.js to have a cartProductSchema
Now user has a cartProductSchema that is defined at the top, and the
user cart now has an array of [cartProductSchema]. Tested that PATCHing
to the User cart still works correctly and it does, now we are also
given on _id for the CartProduct which is exactly what we wanted.
Next Steps: Remove cart product from a user cart using _id.